+++ src/dotnet/APIView/ClientSPA/src/app/_services/reviews/reviews.service.ts
@@ -69,14 +69,6 @@ export class ReviewsService {
     return this.http.get<Review>(this.baseUrl + `/${reviewId}`, { withCredentials: true });
   }
 
-  getAllowedApprovers() : Observable<string[]> {
-    return this.http.get<string[]>(this.baseUrl + `/allowedApprovers`, { withCredentials: true });
-  }
-
-  getPreferredApprovers(reviewId: string) : Observable<string[]> {
-    return this.http.get<string[]>(this.baseUrl + `/${reviewId}/preferredApprovers`, { withCredentials: true });
-  }
-
   getEnableNamespaceReview() : Observable<boolean> {
     return this.http.get<boolean>(this.baseUrl + `/enableNamespaceReview`, { withCredentials: true });
   }